# openapi.yaml
openapi: 3.1.0
info:
  title: Афиша MAX — API
  description: Бэкенд мини-приложения и чат-бота афиши с умным подбором событий.
  version: 1.0.0
paths:
...
  /api/v1/polls:
    post:
      tags:
      - polls
      summary: Create Poll
      description: 'Голосование «куда пойдём?»: ссылку на него можно отправить друзьям.'
      operationId: create_poll_api_v1_polls_post
      parameters:
      - name: x-max-init-data
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: window.WebApp.initData
          title: X-Max-Init-Data
        description: window.WebApp.initData
      - name: x-dev-user-id
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Только при DEV_AUTH_ENABLED
          title: X-Dev-User-Id
        description: Только при DEV_AUTH_ENABLED
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PollCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PollOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - polls
      summary: My Polls
      description: Голосования пользователя (созданные им и те, где он голосовал), новые сверху.
      operationId: my_polls_api_v1_polls_get
      parameters:
      - name: x-max-init-data
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: window.WebApp.initData
          title: X-Max-Init-Data
        description: window.WebApp.initData
      - name: x-dev-user-id
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Только при DEV_AUTH_ENABLED
          title: X-Dev-User-Id
        description: Только при DEV_AUTH_ENABLED
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PollSummaryOut'
                title: Response My Polls Api V1 Polls Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/polls/{poll_id}:
    get:
      tags:
      - polls
      summary: Get Poll
      operationId: get_poll_api_v1_polls__poll_id__get
      parameters:
      - name: poll_id
        in: path
        required: true
        schema:
          type: integer
          title: Poll Id
      - name: x-max-init-data
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: window.WebApp.initData
          title: X-Max-Init-Data
        description: window.WebApp.initData
      - name: x-dev-user-id
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Только при DEV_AUTH_ENABLED
          title: X-Dev-User-Id
        description: Только при DEV_AUTH_ENABLED
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PollOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/polls/{poll_id}/votes:
    post:
      tags:
      - polls
      summary: Vote
      operationId: vote_api_v1_polls__poll_id__votes_post
      parameters:
      - name: poll_id
        in: path
        required: true
        schema:
          type: integer
          title: Poll Id
      - name: x-max-init-data
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: window.WebApp.initData
          title: X-Max-Init-Data
        description: window.WebApp.initData
      - name: x-dev-user-id
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Только при DEV_AUTH_ENABLED
          title: X-Dev-User-Id
        description: Только при DEV_AUTH_ENABLED
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VoteIn'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PollOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    CategoryOut:
      properties:
        id:
          type: integer
          title: Id
        slug:
          type: string
          title: Slug
        title:
          type: string
          title: Title
        emoji:
          type: string
          title: Emoji
      type: object
      required:
      - id
      - slug
      - title
      - emoji
      title: CategoryOut
    CityOut:
      properties:
        id:
          type: integer
          title: Id
        name:
          type: string
          title: Name
      type: object
      required:
      - id
      - name
      title: CityOut
    EventOut:
      properties:
        id:
          type: integer
          title: Id
        title:
          type: string
          title: Title
        description:
          type: string
          title: Description
        category:
          $ref: '#/components/schemas/CategoryOut'
        venue:
          $ref: '#/components/schemas/VenueOut'
        starts_at:
          type: string
          format: date-time
          title: Starts At
        ends_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Ends At
        price_min:
          type: integer
          title: Price Min
        price_max:
          anyOf:
          - type: integer
          - type: 'null'
          title: Price Max
        is_free:
          type: boolean
          title: Is Free
        age_limit:
          type: integer
          title: Age Limit
        pushkin_card:
          type: boolean
          title: Pushkin Card
        tags:
          items:
            type: string
          type: array
          title: Tags
        image_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Image Url
        ticket_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Ticket Url
        source:
          type: string
          title: Source
        source_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Source Url
        status:
          type: string
          title: Status
        is_test_data:
          type: boolean
          title: Is Test Data
        data_updated_at:
          type: string
          format: date-time
          title: Data Updated At
        is_favorite:
          type: boolean
          title: Is Favorite
          default: false
      type: object
      required:
      - id
      - title
      - description
      - category
      - venue
      - starts_at
      - ends_at
      - price_min
      - price_max
      - is_free
      - age_limit
      - pushkin_card
      - tags
      - image_url
      - ticket_url
      - source
      - source_url
      - status
      - is_test_data
      - data_updated_at
      title: EventOut
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    PollCreate:
      properties:
        event_ids:
          items:
            type: integer
          type: array
          maxItems: 5
          minItems: 2
          title: Event Ids
        title:
          anyOf:
          - type: string
            maxLength: 200
          - type: 'null'
          title: Title
      type: object
      required:
      - event_ids
      title: PollCreate
    PollOptionOut:
      properties:
        id:
          type: integer
          title: Id
        event:
          $ref: '#/components/schemas/EventOut'
        votes:
          type: integer
          title: Votes
      type: object
      required:
      - id
      - event
      - votes
      title: PollOptionOut
    PollOut:
      properties:
        id:
          type: integer
          title: Id
        title:
          type: string
          title: Title
        options:
          items:
            $ref: '#/components/schemas/PollOptionOut'
          type: array
          title: Options
        total_votes:
          type: integer
          title: Total Votes
        my_option_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: My Option Id
      type: object
      required:
      - id
      - title
      - options
      - total_votes
      - my_option_id
      title: PollOut
    PollSummaryOut:
      properties:
        id:
          type: integer
          title: Id
        title:
          type: string
          title: Title
        created_at:
          type: string
          format: date-time
          title: Created At
        total_votes:
          type: integer
          title: Total Votes
        options_count:
          type: integer
          title: Options Count
        is_mine:
          type: boolean
          title: Is Mine
          description: Пользователь — автор голосования
        voted:
          type: boolean
          title: Voted
          description: Пользователь уже проголосовал
        events_preview:
          items:
            type: string
          type: array
          title: Events Preview
      type: object
      required:
      - id
      - title
      - created_at
      - total_votes
      - options_count
      - is_mine
      - voted
      - events_preview
      title: PollSummaryOut
      description: 'Голосование в списке: без вариантов, только то, что нужно, чтобы найти нужное.'
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
    VenueOut:
      properties:
        id:
          type: integer
          title: Id
        name:
          type: string
          title: Name
        address:
          type: string
          title: Address
        latitude:
          anyOf:
          - type: number
          - type: 'null'
          title: Latitude
        longitude:
          anyOf:
          - type: number
          - type: 'null'
          title: Longitude
        city:
          $ref: '#/components/schemas/CityOut'
      type: object
      required:
      - id
      - name
      - address
      - latitude
      - longitude
      - city
      title: VenueOut
    VoteIn:
      properties:
        option_id:
          type: integer
          title: Option Id
      type: object
      required:
      - option_id
      title: VoteIn